ENH: Handle PODs that have a special status
When a POD has become evicted, the JSON has a .status like this:
,----[ kubectl get pod -o json ]----
| {
|     "status": {
|         "message": "The node was low on resource: nodefs.",
|         "phase": "Failed",
|         "reason": "Evicted",
|         "startTime": "2018-07-03T21:30:39Z"
|     }
| }
`----

Show the .status.reason in podcontainerages; omit any POD with such a status (assuming it is not healthy) from podcontainernames, podhosts, podimages.
Without excluding those, the output of kubectl-podcontainernames may not be unique, and then commands like kubectl logs will barf on it.