kubectl-wait: When there are multiple Pod results, filter for running ones
To obtain more precise errors, and if there's still (/ again) a single running Pod, just use that one and ignore the other ones (which probably are in state "terminating" or "waiting"). This way, we can execute the SUBCOMMAND earlier; before, we had to retry until the previous Pod was fully removed and not just in state "terminating".