kubectl wrapper: Return the kubectl status (Bash-only) when using pager
set -o pipefail would also work in Korn shell, but as we're in a function the undo would be difficult. Therefore, use Bash's PIPESTATUS (sorry Korn shell; you're not my preferred shell, anyway).
Note that for the pipeline indexing to work when eval is involved, the PIPESTATUS array has to be accessed within the eval; in other words, outside of it, eval counts as a single pipeline already.