kubectl-ef: Use absolute timestamps
The relative time differences that kubectl prints out are not helpful when following along; after the initial backlog of events, all new ones have "0s", and the time passed is lost.
We can obtain the full timestamp via a custom-columns argument. The original OBJECT seems to be a concatenation of (lowercased) KIND + / + OBJECT; kubectl expects jsonpath, which supports a concat(v1, v2) syntax, but it's now accepted. Therefore, we have to split this into two separate columns: KIND and OBJECT.
Adding a reldate filter sounds like a good idea, but it looks like it needs some --unbuffered option first.