Refactoring: Move the implementation back from the kubectl() function to kubectl-wrapper
The downside of kubectl-wrapper delegating to kubectl() is the ugly sourcing of shell/wrappers.sh and that kubectl() is then defined as a function. (But as long as it's not exported, this should be fine.)
Let's do it the other way around; the only thing that the function needs to do is incrementing the _kubectl_use counter in the current shell.