BUG: UI: Shell command-line is overwritten
The first {status,spinning}printf must not reposition the cursor one line above (and clear that line); in an interactive shell, that line contains the command invocation.
parse() is invoked as part of a pipeline; we therefore cannot simply use a global isFirst flag, but need to pass it in.