kubectl-podage: ENH: Allow age filtering
Implementation taken from git-localbr.
One added complication here is that the default output format cannot be extended, and in order to emulate it (using either json, go-template, or the easiest --output=custom-columns), we have to do processing of the raw data, as the relative age and the tally of running containers is not directly accessible, but generated internally.