ENH: Add restart count to output
When the container basically crashes immediately, even the KUBECTL_WAIT_INTERVAL may be too large to intercept the container while it's alive.
This would be hard to recognize, the command would just keep waiting and waiting. With the addition of the restart count, the user at least can recognize that this has happened.